Broken down the stream endpoints function Added broadcaster and ingest endpoints
